[OPF #907] special version instead of sha1
on admin info page

- moved ChiliProject::Info in OpenProject namespace
