+++ app/views/admin/info.html.erb
@@ -15,12 +15,12 @@ See doc/COPYRIGHT.rdoc for more details.
 <p><strong><%= Redmine::Info.versioned_name %></strong> (<%= @db_adapter_name %>)</p>
 
 <table class="list">
-<% @checklist.each do |label, result| %>
-	<tr class="<%= cycle 'odd', 'even' %>">
-		<td><%= l(label) %></td>
-		<td width="30px"><%= image_tag((result ? 'check.png' : 'error.png'), :style => "vertical-align:bottom;", :alt => (result ? l(:general_text_Yes) : l(:general_text_No))) %></td>
-	</tr>
-<% end %>
+  <% @checklist.each do |label, result| %>
+    <tr class="<%= cycle 'odd', 'even' %>">
+      <td><%= l(label) %></td>
+      <td width="30px"><%= image_tag((result ? 'check.png' : 'error.png'), :style => "vertical-align:bottom;", :alt => (result ? l(:general_text_Yes) : l(:general_text_No))) %></td>
+    </tr>
+  <% end %>
 </table>
 
 <% html_title(l(:label_information_plural)) -%>
